# docs/openapi/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "AI Agent Productivity Platform - Open API",
        "contact": {},
        "version": "v0.0.1"
    },
    "basePath": "/openapi/v1",
    "paths": {
...
        "/file/upload/direct": {
            "post": {
                "security": [
                    {
                        "JWT": []
                    }
                ],
                "description": "文件上传",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "openapi"
                ],
                "summary": "文件上传",
                "parameters": [
                    {
                        "type": "string",
                        "description": "原始文件名",
                        "name": "fileName",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "文件",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/response.UploadFileResp"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "response.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "msg": {
                    "type": "string"
                }
            }
        },
        "response.UploadFileResp": {
            "type": "object",
            "properties": {
                "status": {
                    "description": "0:上传失败，1：上传成功",
                    "type": "integer"
                }
            }
        }
    }
}